+++ package.json
@@ -2,7 +2,7 @@
     "name": "catboys",
     "version": "1.0.5",
     "description": "JS wrapper for the Catboys API",
-    "main": "index.js",
+    "main": "src/index.js",
     "typings": "./typings/index.d.ts",
     "scripts": {
         "format": "prettier -w ."
@@ -15,7 +15,7 @@
         "catboys",
         "catboys.com"
     ],
-    "author": "Kristian Kramer",
+    "author": "Catboys.com",
     "license": "MIT",
     "bugs": {
         "url": "https://github.com/Catboys-Dev/catboys-js/issues"
