+++ typings/index.d.ts
@@ -1,4 +1,4 @@
-//List all the functions
+// List all the functions
 declare class CatboyClient {
     sfw: {
         img(): Promise<CatboyClient.CatboyRequestResults>;
@@ -9,21 +9,24 @@ declare class CatboyClient {
     };
 }
 
-export = CatboyClient;
-
 declare namespace CatboyClient {
-    //Help create options interface for the few functions that need it
+    // Help create options interface for the few functions that need it
     export interface CatboyQueryParams {
         text: string;
     }
+
     export interface CatboyRequestResults {
         url: string;
     }
+
     export interface CatboyChatResults {
         response: string;
         url?: string;
     }
+
     export interface CatboyCatResult {
         catboy: string;
     }
 }
+
+export = CatboyClient;
